Feat : 기능 추가 - 배포

+++ GodSaeng/src/main/java/com/planner/godsaeng/dto/ChallengeStatusDTO.java
@@ -26,7 +26,7 @@ public class ChallengeStatusDTO {
     private Date cstartdate;
     private Date cenddate;
     private Integer ctypeofverify;
-    private BigInteger datediff;
+    private Integer datediff;
     private BigInteger totalcount;
     private BigInteger cvsuccesscount;
