+++ GodSaeng/src/main/java/com/planner/godsaeng/service/ChallengeDataConverter.java
@@ -20,7 +20,7 @@ public class ChallengeDataConverter implements Converter<Object[], ChallengeStat
         convertedChallengeStatusDTO.setCstartdate((Date) source[3]);
         convertedChallengeStatusDTO.setCenddate((Date) source[4]);
         convertedChallengeStatusDTO.setCtypeofverify((Integer) source[5]);
-        convertedChallengeStatusDTO.setDatediff((BigInteger) source[6]);
+        convertedChallengeStatusDTO.setDatediff((Integer) source[6]);
         convertedChallengeStatusDTO.setTotalcount((BigInteger) source[7]);
         convertedChallengeStatusDTO.setCvsuccesscount(((BigInteger) source[9]));
         return convertedChallengeStatusDTO;
